+++ birdo/src/birdo/utilities/feather.java
@@ -7,7 +7,7 @@ public class feather extends object {
 	public int speed;
 	public player p;
 
-	public feather(int x, int y, boolean forward) {
+	public feather(double x, double y, boolean forward) {
 		super(x, y, 12, 8, Color.BLACK);
 		this.forward = forward;
